+++ CHANGELOG.md
@@ -7,13 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-
-## [0.0.2] - 2020-03-06
+## [0.2.0] - 2020-04-03
 
 ### Added
 
-- Test TilePile class
-- Test PriceChart class
+- Tested Bank class
+- Tested StockSet class
+- updated README.md
+- Created GameEngine class
+- Tested CellLocation class
+- Tested Player class
+- Tested HotelChain class
+
+## [0.1.0] - 2020-03-06
+
+### Added
+
+- Tested TilePile class
+- Tested PriceChart class
 - Updated README.md
 - Made changes to bank
 - Added TurnManagerTest
@@ -30,8 +41,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[unreleased]: https://github.com/cs3321isu/2263-S2020-Team-2/compare/v0.0.2...HEAD
+[unreleased]: https://github.com/cs3321isu/2263-S2020-Team-2/compare/v0.2.0...HEAD
 
-[0.0.2]: https://github.com/cs3321isu/2263-S2020-Team-2/compare/v0.0.1...v0.0.2
+[0.2.0]: https://github.com/cs3321isu/2263-S2020-Team-2/compare/v0.0.1...v0.2.0
+[0.1.0]: https://github.com/cs3321isu/2263-S2020-Team-2/compare/v0.0.1...v0.1.0
 [0.0.1]: https://github.com/cs3321isu/2263-S2020-Team-2/releases/tag/v0.0.1